Changed FLIPENDIAN_INT16() to properly flip signed integers.

+++ mess/utils.h
@@ -167,7 +167,7 @@ char *strip_extension(const char *filename);
 int compute_log2(int val);
 
 /* Endian macros */
-#define FLIPENDIAN_INT16(x)	((((x) >> 8) | ((x) << 8)) & 0xffff)
+#define FLIPENDIAN_INT16(x)	(((((UINT16) (x)) >> 8) | ((x) << 8)) & 0xffff) 
 #define FLIPENDIAN_INT32(x)	((((x) << 24) | (((UINT32) (x)) >> 24) | \
                        (( (x) & 0x0000ff00) << 8) | (( (x) & 0x00ff0000) >> 8)))
